pothole-detection/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
numpy==1.26.4
pydantic==2.6.1
tensorflow==2.15.0
geopy==2.4.1
python-dotenv==1.0.1
